Add GetFranchiseSeasonBattingStatistics and GetFranchiseSeasonPitchingStatistics stubs
